fix: triple-email-template elements test code 분리
